Create reviewer/images.ts
